+++ camera_info_manager_metadata_extractor/CHANGELOG.rst
@@ -5,6 +5,15 @@
 Changelog for package camera_info_manager_metadata_extractor
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+3.0.1 (2025-05-13)
+------------------
+
+3.0.0 (2025-05-13)
+------------------
+* Finished the refactor for timed metadata support.
+* Big refactoring towards better handling of timed metadata.
+* Contributors: Martin Pecka
+
 2.0.3 (2025-02-27)
 ------------------